--- force-app/main/default/lwc/utils/utils.html
@@ -1,3 +1,0 @@
-<template>
-    
-</template>
